Second Change

+++ src/processor.cpp
@@ -5,16 +5,16 @@
 float Processor::Utilization() {
   std::vector<std::string> value = LinuxParser::CpuUtilization();
 
-  float user = std::stof(value[LinuxParser::kUser_]);
-  float nice = std::stof(value[LinuxParser::kNice_]);
-  float system = std::stof(value[LinuxParser::kSystem_]);
-  float idle = std::stof(value[LinuxParser::kIdle_]);
-  float iowait = std::stof(value[LinuxParser::kIOwait_]);
-  float irq = std::stof(value[LinuxParser::kIRQ_]);
-  float softirq = std::stof(value[LinuxParser::kSoftIRQ_]);
-  float steal = std::stof(value[LinuxParser::kSteal_]);
-  float guest = std::stof(value[LinuxParser::kGuest_]);
-  float guest_nice = std::stof(value[LinuxParser::kGuestNice_]);
+  double user = std::stof(value[LinuxParser::kUser_]);
+  double nice = std::stof(value[LinuxParser::kNice_]);
+  double system = std::stof(value[LinuxParser::kSystem_]);
+  double idle = std::stof(value[LinuxParser::kIdle_]);
+  double iowait = std::stof(value[LinuxParser::kIOwait_]);
+  double irq = std::stof(value[LinuxParser::kIRQ_]);
+  double softirq = std::stof(value[LinuxParser::kSoftIRQ_]);
+  double steal = std::stof(value[LinuxParser::kSteal_]);
+  double guest = std::stof(value[LinuxParser::kGuest_]);
+  double guest_nice = std::stof(value[LinuxParser::kGuestNice_]);
 
   float Total_CPU_time_since_boot =
       user + nice + system + idle + iowait + irq + softirq + steal;
